pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "dotenv-encrypt"
version = "0.1.0"
description = "Encrypt, decrypt, and load .env files with authenticated AES-256-GCM encryption."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "swiftbird07" }]
keywords = ["dotenv", "env", "encryption", "secrets", "aes-gcm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "cryptography>=41",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "build>=1.2",
    "pytest>=8",
    "ruff>=0.5",
    "twine>=5",
]

[project.urls]
Homepage = "https://github.com/swiftbird07/dotenv-encrypt"
Repository = "https://github.com/swiftbird07/dotenv-encrypt"
Issues = "https://github.com/swiftbird07/dotenv-encrypt/issues"

[project.scripts]
dotenv-encrypt = "dotenv_encrypt.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/dotenv_encrypt"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "S"]
ignore = [
    "S101",  # pytest assertions are fine in tests
]
